+++ NEWS.md
@@ -12,6 +12,11 @@ faixas serao obtidos segundo as regras:
     * minimo: o maior multiplo de `byX` menor que min(X)
     * maximo: o menor multiplo de `byX` maior que max(X)
 
+# master (dev)
+
+* `plot.interpolador` agora usa `...` para passar argumentos a `geragrade`, permitindo maior 
+  controle sobre a superficie gerada
+
 # curvacolina 1.0.1
 
 ## Bug fixes